Move create_fs(), set_uuid and xfs_growfs() to util.rs
create_fs() is shared by MDV and filesystem - moving to a more suitable
place.  In the future, we will want to grow the MDV, so xfs_growfs moved
also.
